+++ core/src/main/java/io/token/tokenrequest/TokenRequest.java
@@ -445,6 +445,20 @@ public abstract class TokenRequest {
         }
 
         /**
+         * Optional. In the scenario where TPP wishes to know the user's selection of country and
+         * bank, TPP should provide this url so that Token can make a call with relevant
+         * information as parameters. TPP can use that information to set transfer destination.
+         *
+         * @param url URL
+         * @return builder
+         */
+        public TransferBuilder setSetTransferDestinationsUrl(String url) {
+            this.requestPayload.getTransferBodyBuilder()
+                    .setSetTransferDestinationsUrl(url);
+            return this;
+        }
+          
+        /**
          * Optional. Sets the ultimate party to which the money is due.
          *
          * @param ultimateCreditor the ultimate creditor